PLT-8058 - Abstracted domain data generators into their own file
